Move formatters/destinations to source files
