--- libs/logging/include/hpx/logging/format.hpp
@@ -1,26 +1,0 @@
-// format.hpp
-
-// Boost Logging library
-//
-// Author: John Torjo, www.torjo.com
-//
-// Copyright (C) 2007 John Torjo (see www.torjo.com for email)
-//
-//  SPDX-License-Identifier: BSL-1.0
-// Distributed under the Boost Software License, Version 1.0.
-//    (See accompanying file LICENSE_1_0.txt or copy at
-//          http://www.boost.org/LICENSE_1_0.txt)
-//
-// See http://www.boost.org for updates, documentation, and revision history.
-// See http://www.torjo.com/log2/ for more details
-
-// this is fixed!
-#ifndef JT28092007_format_HPP_DEFINED
-#define JT28092007_format_HPP_DEFINED
-
-#include <hpx/config.hpp>
-#include <hpx/logging/format/destination/defaults.hpp>
-#include <hpx/logging/format/formatter/defaults.hpp>
-#include <hpx/logging/manipulator.hpp>
-
-#endif
